Añadidas peticiones de confirmacion al borrar producto o socio

+++ ProyectoFinal/ProyectoFinal/ElimProductoForm.vb
@@ -121,21 +121,26 @@ Public Class ElimProductoForm
 
     Private Sub ButtonEliminar_Click(sender As Object, e As EventArgs) Handles ButtonEliminar.Click
         Dim codigo = TextBoxCodProd.Text
-        'Eliminamos la fila del dataset
-        ds.Tables("PRODUCTO").Rows(contador).Delete()
-        'Hay que eliminar la fila del dataAdapter!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-        'Limpiamos todos los textBox de la fila
-        limpiarTextBox()
+        If (MessageBox.Show("¿Estás seguro de que quieres eliminar este registro?", "Eliminar", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) =
+         DialogResult.No) Then
+            MsgBox("Operación cancelada")
+        Else
+            'Eliminamos la fila del dataset
+            ds.Tables("PRODUCTO").Rows(contador).Delete()
+            'Hay que eliminar la fila del dataAdapter!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+            'Limpiamos todos los textBox de la fila
+            limpiarTextBox()
 
 
-        ' consulta sql para eliminar de la bd
-        da.DeleteCommand = New SqlCommand("DELETE FROM PRODUCTO WHERE cod_prod=@cod_prod", con)
-        da.DeleteCommand.Parameters.Add("@cod_prod", SqlDbType.Int).Value = codigo
+            ' consulta sql para eliminar de la bd
+            da.DeleteCommand = New SqlCommand("DELETE FROM PRODUCTO WHERE cod_prod=@cod_prod", con)
+            da.DeleteCommand.Parameters.Add("@cod_prod", SqlDbType.Int).Value = codigo
 
-        ' update
-        da.Update(ds, "PRODUCTO")
-        MsgBox("ELIMINADO")
-
+            ' update
+            da.Update(ds, "PRODUCTO")
+            MsgBox("ELIMINADO")
+        End If
     End Sub
 End Class
